@@ -1,6 +1,4 @@
 class ListsWord < ActiveRecord::Base
-
-	belongs_to :list
-	belongs_to :word
-
+  belongs_to :list
+  belongs_to :word
 end
